ported r34862:35361 from main
git-svn-id: https://svn.r-project.org/R/branches/R-noNULLenv@35372 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/stats/R/loess.R
@@ -178,7 +178,7 @@ simpleLoess <-
 		as.integer(nonparametric),
 		as.integer(order.drop.sqr),
 		as.integer(sum.drop.sqr),
-		as.integer(span*cell),
+		as.double(span*cell),
 		as.character(surf.stat),
 		temp = double(N),
 		parameter = integer(7),
